change killeveryon comp and system to grantobjectives
